refactor(template): 调整直方图观察方法的参数类型为int64
将Observe方法的观测值参数从float64修改为int64，统一指标数值类型规范

+++ internal/template/template.go
@@ -129,7 +129,7 @@ func (m *{{ $metricName }}Metric) Dec({{ $metric.Labels | labelParams }}) {
 {{- if eq $method "observe" }}
 
 // Observe 观察直方图值
-func (m *{{ $metricName }}Metric) Observe(v float64, {{ $metric.Labels | labelParams }}) {
+func (m *{{ $metricName }}Metric) Observe(v int64, {{ $metric.Labels | labelParams }}) {
 	m.vec.Observe(v, {{ $metric.Labels | labelArgs }})
 }
 {{- end }}
